requirements.txt
pymysql
djangorestframework
django-cors-headers
